Put test into a designated file
